+++ Plugins/ActorPools/Source/NexusActorPools/Public/NGetActorBlueprintAsyncAction.h
@@ -36,6 +36,7 @@ public:
 	FGetActorAsyncOutputPin Completed;
 
 	virtual void Activate() override;
+	virtual void BeginDestroy() override;
 
 private:
 	FDelegateHandle OnCreatedPoolHandle;