+++ Plugins/ActorPools/Source/NexusActorPools/Public/NSpawnActorBlueprintAsyncAction.h
@@ -38,6 +38,7 @@ public:
 	FSpawnActorAsyncOutputPin Completed;
 
 	virtual void Activate() override;
+	virtual void BeginDestroy() override;
 
 private:
 	FDelegateHandle OnCreatedPoolHandle;