+++ Plugins/ProcGen/Source/NexusProcGen/Public/Cell/NCellVoxelData.h
@@ -15,7 +15,7 @@ UENUM(meta=(Bitflags,UseEnumValuesAsMaskValuesInEditor=true))
 enum class ENCellVoxel : uint8
 {
 	Empty		= 0,
-	Occupied	= 1 << 0
+	Occupied	= 1
 };
 ENUM_CLASS_FLAGS(ENCellVoxel)